admin probe done, probing works, needs stats
