+++ sw/admin-cli/fpga-stats.hh
@@ -1,0 +1,3 @@
+#pragma once
+
+stats_t s;
